requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.43.0
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-auth==0.1.9
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-rest-auth==0.9.5
djangorestframework==3.16.0
idna==3.10
pillow==11.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.3.0
